Clean up nuxt config
