+++ nuxt.config.ts
@@ -1,6 +1,5 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
-import fs from "node:fs/promises"
-import path from "node:path"
+import { copyPrismaAssets } from "./server/utils/copy-prisma-assets"
 
 export default defineNuxtConfig({
   modules: [
@@ -61,22 +60,7 @@ export default defineNuxtConfig({
   },
 
   hooks: {
-    "nitro:build:public-assets": async (nitro) => {
-      // Copy Prisma schema and migrations to the output directory
-      const prismaInDir = path.join(__dirname, "prisma")
-      const prismaOutDir = path.join(nitro.options.output.dir, "prisma")
-      await fs.cp(prismaInDir, prismaOutDir, { recursive: true })
-
-      // Copy Prisma schema engine to the output directory
-      const prismaEnginesInDir = path.join(__dirname, "node_modules", "@prisma", "engines")
-      const prismaEnginesOutDir = path.join(nitro.options.output.serverDir, "node_modules", "@prisma", "engines")
-      const prismaEnginesFiles = await fs.readdir(prismaEnginesInDir)
-      for (const file of prismaEnginesFiles) {
-        if (/^schema-engine-.+$/.test(file)) {
-          await fs.cp(path.join(prismaEnginesInDir, file), path.join(prismaEnginesOutDir, file))
-        }
-      }
-    },
+    "nitro:build:public-assets": copyPrismaAssets,
   },
 
   eslint: {
